requirements.txt
requests
tinydb
python-dotenv
aiohttp